siliconflow/requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
SpeechRecognition>=3.8.1
pydub>=0.25.1
numpy>=1.22.0
